+++ packages/plugin-link-card/src/constants/icon-svg.ts
@@ -1,0 +1,14 @@
+/**
+ * @description icon svg
+ * @author wangfupeng
+ */
+
+/**
+ * 【注意】svg 字符串的长度 ，否则会导致代码体积过大
+ * 尽量选择 https://www.iconfont.cn/collections/detail?spm=a313x.7781069.0.da5a778a4&cid=20293
+ * 找不到再从 iconfont.com 搜索
+ */
+
+// // 垃圾桶（删除）
+// export const TRASH_SVG =
+//   '<svg viewBox="0 0 1024 1024"><path d="M128 320v640c0 35.2 28.8 64 64 64h576c35.2 0 64-28.8 64-64V320H128z m192 576H256V448h64v448z m128 0h-64V448h64v448z m128 0h-64V448h64v448z m128 0h-64V448h64v448zM848 128H640V48c0-26.4-21.6-48-48-48h-224c-26.4 0-48 21.6-48 48V128H112C85.6 128 64 149.6 64 176V256h832V176c0-26.4-21.6-48-48-48zM576 128h-192V64.8h192V128z"></path></svg>'