+++ packages/plugin-link-card/src/index.ts
@@ -1,0 +1,12 @@
+/**
+ * @description src entry
+ * @author wangfupeng
+ */
+
+import './asserts/index.less'
+
+import module from './module/index'
+
+export * from './module/custom-types'
+
+export default module