+++ packages/basic-modules/CHANGELOG.md
@@ -1,5 +1,40 @@
 # Change Log
 
+## 1.5.28
+
+### Patch Changes
+
+- Updated dependencies [7225111]
+  - @wangeditor-next/core@1.7.28
+
+## 1.5.27
+
+### Patch Changes
+
+- Updated dependencies [2099e64]
+  - @wangeditor-next/core@1.7.27
+
+## 1.5.26
+
+### Patch Changes
+
+- 8c762a9: feat(code block): add code block handletab
+
+## 1.5.25
+
+### Patch Changes
+
+- Updated dependencies [5fa2a72]
+  - @wangeditor-next/core@1.7.26
+
+## 1.5.24
+
+### Patch Changes
+
+- 5bbff78: fix: add more font style
+- Updated dependencies [7b80f5d]
+  - @wangeditor-next/core@1.7.25
+
 ## 1.5.23
 
 ### Patch Changes