+++ packages/code-highlight/CHANGELOG.md
@@ -1,5 +1,33 @@
 # Change Log
 
+## 1.3.25
+
+### Patch Changes
+
+- Updated dependencies [7225111]
+  - @wangeditor-next/core@1.7.28
+
+## 1.3.24
+
+### Patch Changes
+
+- Updated dependencies [2099e64]
+  - @wangeditor-next/core@1.7.27
+
+## 1.3.23
+
+### Patch Changes
+
+- Updated dependencies [5fa2a72]
+  - @wangeditor-next/core@1.7.26
+
+## 1.3.22
+
+### Patch Changes
+
+- Updated dependencies [7b80f5d]
+  - @wangeditor-next/core@1.7.25
+
 ## 1.3.21
 
 ### Patch Changes