+++ packages/plugin-formula/CHANGELOG.md
@@ -1,5 +1,74 @@
 # @wangeditor-next/plugin-formula
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [7225111]
+  - @wangeditor-next/editor@5.6.22
+
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [2099e64]
+  - @wangeditor-next/editor@5.6.21
+
+## 1.0.5
+
+### Patch Changes
+
+- Updated dependencies [8c762a9]
+  - @wangeditor-next/editor@5.6.20
+
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [5fa2a72]
+  - @wangeditor-next/editor@5.6.19
+
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [ad3d49b]
+  - @wangeditor-next/editor@5.6.18
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [230e46d]
+  - @wangeditor-next/editor@5.6.17
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [33d87c9]
+  - @wangeditor-next/editor@5.6.16
+
+## 1.0.0
+
+### Major Changes
+
+- 7b050b7: publish first major version
+
+## 0.1.0
+
+### Minor Changes
+
+- 5b24c79: fix(formula): add mathml output config
+
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [5bbff78]
+- Updated dependencies [7b80f5d]
+  - @wangeditor-next/editor@5.6.15
+
 ## 0.0.10
 
 ### Patch Changes