+++ packages/video-module/CHANGELOG.md
@@ -1,5 +1,39 @@
 # Change Log
 
+## 1.3.34
+
+### Patch Changes
+
+- Updated dependencies [7225111]
+  - @wangeditor-next/core@1.7.28
+
+## 1.3.33
+
+### Patch Changes
+
+- Updated dependencies [2099e64]
+  - @wangeditor-next/core@1.7.27
+
+## 1.3.32
+
+### Patch Changes
+
+- Updated dependencies [5fa2a72]
+  - @wangeditor-next/core@1.7.26
+
+## 1.3.31
+
+### Patch Changes
+
+- 33d87c9: video module add align prop
+
+## 1.3.30
+
+### Patch Changes
+
+- Updated dependencies [7b80f5d]
+  - @wangeditor-next/core@1.7.25
+
 ## 1.3.29
 
 ### Patch Changes